Native x64 support
Implemented the changes in issue #19 introducing native x64 support. This
means that x64 builds can only be done on x64 computers.

The make_install.cmd has been changed to take two arguments. Run it without
arguments for an explanation.

The patch is pretty extensive as some code layout changes was done as well.