Fixes in wpkg-gp-*.dll/exe
Removed making wpkg-gp trying to start the service, as it is not necessary
to start it, only to wait.

Hopefully this will make debugging easier, as well as resolve issue #22 as
it seems that wpkg-gp was always waiting for 2 minutes on the service
if it wasn't started yet, no matter what. This fix will return immediately
when the service is started.